Create Squadron.java
Add fields corresponding to the lines in the text file.

These are all private because the squadron shouldn't expose its private data to other game Components.  The other items will query it for its data.  This is to make it harder for outside objects to modify the state of this object.