Create very basic file parser
Basically this will read things in line by line.  If the line isn't empty it will create a key value pair and populate a field.

Every non-empty line it will check if all fields are populated and then if so, will build a new squadron object and put it into the hashmap.  Later when we need a squad, we will grab a copy of it out of the hashmap.