Initial path setup for constructing the initial attack pool
The source of the attack pool should decide which dice in its armament are appropriate and create them, and then pass them into the constructor.

This is because the dice don't always respect the ranges. in particular squadron dice don't don't care about ranges.  A few others might as well.