Create a dice builder function
Lets make this static so that we don't need to instantiate an object to use it.  This is more of a utility function intended to be used to create the initial pool or add dice etc.

Not wild about creating the defs over and over with each call, might break that out to a static field later since they don't change